requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.4.2
jinja2==3.1.2
aiofiles==23.2.1
websockets==12.0

# Model Management
torch==2.1.1
transformers==4.35.2
accelerate==0.24.1
sentencepiece==0.1.99
protobuf==4.25.1
ollama==0.1.6

# Knowledge Base
chromadb==0.4.18
langchain==0.0.350
python-docx==1.0.1
unstructured==0.10.30
markdown==3.5.2
beautifulsoup4==4.12.2

# System Monitoring
psutil==5.9.6
gputil==1.4.0
nvidia-ml-py3==7.352.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0
aiohttp==3.9.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
pytest-cov==4.1.0
black==23.10.1
flake8==6.1.0
mypy==1.6.1
sphinx==7.2.6
sphinx-rtd-theme==1.3.0 